# my-prop/src/api/generated/model/website.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * myprop.live API
 * Schema-first contract for the myprop.live app. Until a backend exists it is
 * served by MSW (src/mocks). Paths are relative to the /api server.
 *
 * OpenAPI spec version: 0.1.0
 */
import type { WebsiteStatus } from "./websiteStatus.ts"

export interface Website {
  id: string
  name: string
  /** The website's myprop.live address. */
  domain: string
  status: WebsiteStatus
  thumbnailUrl: string
  /** @minimum 0 */
  views: number
  /** @minimum 0 */
  leads: number
  /** Leads per 100 views. */
  conversionRate: number
  toolIds: string[]
  templateId: string
  createdAt: string
  updatedAt: string
  /** @nullable */
  publishedAt: string | null
}