# my-prop/src/api/generated/model/websiteContent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * myprop.live API
 * Schema-first contract for the myprop.live app. Until a backend exists it is
 * served by MSW (src/mocks). Paths are relative to the /api server.
 *
 * OpenAPI spec version: 0.1.0
 */
import type { WebsiteInfo } from "./websiteInfo.ts"
import type { WebsiteSection } from "./websiteSection.ts"

export interface WebsiteContent {
  websiteId: string
  info: WebsiteInfo
  /**
   * null until the builder saves for the first time; the builder then starts from its default layout.
   * @nullable
   */
  sections: WebsiteSection[] | null
  /** @nullable */
  savedAt: string | null
}