# my-prop/src/api/generated/model/websiteFile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * myprop.live API
 * Schema-first contract for the myprop.live app. Until a backend exists it is
 * served by MSW (src/mocks). Paths are relative to the /api server.
 *
 * OpenAPI spec version: 0.1.0
 */

/**
 * Metadata of a file attached to the website (uploads aren't implemented yet).
 */
export interface WebsiteFile {
  id: string
  name: string
  /** @minimum 0 */
  size: number
  /** MIME type */
  type: string
}