# my-prop/src/api/generated/model/websiteInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * myprop.live API
 * Schema-first contract for the myprop.live app. Until a backend exists it is
 * served by MSW (src/mocks). Paths are relative to the /api server.
 *
 * OpenAPI spec version: 0.1.0
 */
import type { WebsiteAdditionalContent } from "./websiteAdditionalContent.ts"
import type { WebsiteFile } from "./websiteFile.ts"

/**
 * Project information the website is generated from.
 */
export interface WebsiteInfo {
  /** @minLength 1 */
  projectName: string
  location: string
  description: string
  propertyType: string
  configurations: string[]
  priceRange: string
  amenities: string[]
  targetAudience: string
  aiTone: string
  generateWithAI: boolean
  uploadedFiles: WebsiteFile[]
  additionalContent: WebsiteAdditionalContent
  /** Ids from the website tools catalog. */
  enabledToolIds: string[]
}