# my-prop/src/api/generated/model/websiteInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * myprop.live API
 * Schema-first contract for the myprop.live app. Until a backend exists it is
 * served by MSW (src/mocks). Paths are relative to the /api server.
 *
 * OpenAPI spec version: 0.1.0
 */
import type { WebsiteInfo } from "./websiteInfo.ts"

export interface WebsiteInput {
  templateId: string
  info: WebsiteInfo
}