# my-prop/src/api/generated/model/websiteSection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * myprop.live API
 * Schema-first contract for the myprop.live app. Until a backend exists it is
 * served by MSW (src/mocks). Paths are relative to the /api server.
 *
 * OpenAPI spec version: 0.1.0
 */

/**
 * One page section of the builder document. Its shape belongs to the
 * builder (src/components/builder/types.ts), so the API stores it as-is.
 */
export interface WebsiteSection {
  [key: string]: unknown
}