# my-prop/src/api/generated/websites/websites.msw.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * myprop.live API
 * Schema-first contract for the myprop.live app. Until a backend exists it is
 * served by MSW (src/mocks). Paths are relative to the /api server.
 *
 * OpenAPI spec version: 0.1.0
 */
import { faker } from "@faker-js/faker"

import { HttpResponse, http } from "msw"
import type { RequestHandlerOptions } from "msw"

import type { Website, WebsiteContent } from "../model"

export const getListWebsitesResponseMock = (): Website[] =>
  Array.from(
    { length: faker.number.int({ min: 1, max: 10 }) },
    (_, i) => i + 1
  ).map(() => ({
    id: faker.string.uuid(),
    name: faker.string.alpha({ length: { min: 10, max: 20 } }),
    domain: faker.string.alpha({ length: { min: 10, max: 20 } }),
    status: faker.helpers.arrayElement(["live", "draft"] as const),
    thumbnailUrl: faker.internet.url(),
    views: faker.number.int({ min: 0 }),
    leads: faker.number.int({ min: 0 }),
    conversionRate: faker.number.float({ fractionDigits: 2 }),
    toolIds: Array.from(
      { length: faker.number.int({ min: 1, max: 10 }) },
      (_, i) => i + 1
    ).map(() => faker.string.alpha({ length: { min: 10, max: 20 } })),
    templateId: faker.string.alpha({ length: { min: 10, max: 20 } }),
    createdAt: faker.date.past().toISOString().slice(0, 19) + "Z",
    updatedAt: faker.date.past().toISOString().slice(0, 19) + "Z",
    publishedAt: faker.helpers.arrayElement([
      faker.date.past().toISOString().slice(0, 19) + "Z",
      null,
    ]),
  }))

export const getCreateWebsiteResponseMock = (
  overrideResponse: Partial<Extract<Website, object>> = {}
): Website => ({
  id: faker.string.uuid(),
  name: faker.string.alpha({ length: { min: 10, max: 20 } }),
  domain: faker.string.alpha({ length: { min: 10, max: 20 } }),
  status: faker.helpers.arrayElement(["live", "draft"] as const),
  thumbnailUrl: faker.internet.url(),
  views: faker.number.int({ min: 0 }),
  leads: faker.number.int({ min: 0 }),
  conversionRate: faker.number.float({ fractionDigits: 2 }),
  toolIds: Array.from(
    { length: faker.number.int({ min: 1, max: 10 }) },
    (_, i) => i + 1
  ).map(() => faker.string.alpha({ length: { min: 10, max: 20 } })),
  templateId: faker.string.alpha({ length: { min: 10, max: 20 } }),
  createdAt: faker.date.past().toISOString().slice(0, 19) + "Z",
  updatedAt: faker.date.past().toISOString().slice(0, 19) + "Z",
  publishedAt: faker.helpers.arrayElement([
    faker.date.past().toISOString().slice(0, 19) + "Z",
    null,
  ]),
  ...overrideResponse,
})

export const getGetWebsiteResponseMock = (
  overrideResponse: Partial<Extract<Website, object>> = {}
): Website => ({
  id: faker.string.uuid(),
  name: faker.string.alpha({ length: { min: 10, max: 20 } }),
  domain: faker.string.alpha({ length: { min: 10, max: 20 } }),
  status: faker.helpers.arrayElement(["live", "draft"] as const),
  thumbnailUrl: faker.internet.url(),
  views: faker.number.int({ min: 0 }),
  leads: faker.number.int({ min: 0 }),
  conversionRate: faker.number.float({ fractionDigits: 2 }),
  toolIds: Array.from(
    { length: faker.number.int({ min: 1, max: 10 }) },
    (_, i) => i + 1
  ).map(() => faker.string.alpha({ length: { min: 10, max: 20 } })),
  templateId: faker.string.alpha({ length: { min: 10, max: 20 } }),
  createdAt: faker.date.past().toISOString().slice(0, 19) + "Z",
  updatedAt: faker.date.past().toISOString().slice(0, 19) + "Z",
  publishedAt: faker.helpers.arrayElement([
    faker.date.past().toISOString().slice(0, 19) + "Z",
    null,
  ]),
  ...overrideResponse,
})

export const getGetWebsiteContentResponseMock = (
  overrideResponse: Partial<Extract<WebsiteContent, object>> = {}
): WebsiteContent => ({
  websiteId: faker.string.uuid(),
  info: {
    projectName: faker.string.alpha({ length: { min: 1, max: 20 } }),
    location: faker.string.alpha({ length: { min: 10, max: 20 } }),
    description: faker.string.alpha({ length: { min: 10, max: 20 } }),
    propertyType: faker.string.alpha({ length: { min: 10, max: 20 } }),
    configurations: Array.from(
      { length: faker.number.int({ min: 1, max: 10 }) },
      (_, i) => i + 1
    ).map(() => faker.string.alpha({ length: { min: 10, max: 20 } })),
    priceRange: faker.string.alpha({ length: { min: 10, max: 20 } }),
    amenities: Array.from(
      { length: faker.number.int({ min: 1, max: 10 }) },
      (_, i) => i + 1
    ).map(() => faker.string.alpha({ length: { min: 10, max: 20 } })),
    targetAudience: faker.string.alpha({ length: { min: 10, max: 20 } }),
    aiTone: faker.string.alpha({ length: { min: 10, max: 20 } }),
    generateWithAI: faker.datatype.boolean(),
    uploadedFiles: Array.from(
      { length: faker.number.int({ min: 1, max: 10 }) },
      (_, i) => i + 1
    ).map(() => ({
      id: faker.string.alpha({ length: { min: 10, max: 20 } }),
      name: faker.string.alpha({ length: { min: 10, max: 20 } }),
      size: faker.number.int({ min: 0 }),
      type: faker.string.alpha({ length: { min: 10, max: 20 } }),
    })),
    additionalContent: {
      keyHighlights: faker.string.alpha({ length: { min: 10, max: 20 } }),
      developerInfo: faker.string.alpha({ length: { min: 10, max: 20 } }),
      nearbyLocations: faker.string.alpha({ length: { min: 10, max: 20 } }),
      specialOffers: faker.string.alpha({ length: { min: 10, max: 20 } }),
    },
    enabledToolIds: Array.from(
      { length: faker.number.int({ min: 1, max: 10 }) },
      (_, i) => i + 1
    ).map(() => faker.string.alpha({ length: { min: 10, max: 20 } })),
  },
  sections: faker.helpers.arrayElement([
    Array.from(
      { length: faker.number.int({ min: 1, max: 10 }) },
      (_, i) => i + 1
    ).map(() => ({})),
    null,
  ]),
  savedAt: faker.helpers.arrayElement([
    faker.date.past().toISOString().slice(0, 19) + "Z",
    null,
  ]),
  ...overrideResponse,
})

export const getSaveWebsiteContentResponseMock = (
  overrideResponse: Partial<Extract<WebsiteContent, object>> = {}
): WebsiteContent => ({
  websiteId: faker.string.uuid(),
  info: {
    projectName: faker.string.alpha({ length: { min: 1, max: 20 } }),
    location: faker.string.alpha({ length: { min: 10, max: 20 } }),
    description: faker.string.alpha({ length: { min: 10, max: 20 } }),
    propertyType: faker.string.alpha({ length: { min: 10, max: 20 } }),
    configurations: Array.from(
      { length: faker.number.int({ min: 1, max: 10 }) },
      (_, i) => i + 1
    ).map(() => faker.string.alpha({ length: { min: 10, max: 20 } })),
    priceRange: faker.string.alpha({ length: { min: 10, max: 20 } }),
    amenities: Array.from(
      { length: faker.number.int({ min: 1, max: 10 }) },
      (_, i) => i + 1
    ).map(() => faker.string.alpha({ length: { min: 10, max: 20 } })),
    targetAudience: faker.string.alpha({ length: { min: 10, max: 20 } }),
    aiTone: faker.string.alpha({ length: { min: 10, max: 20 } }),
    generateWithAI: faker.datatype.boolean(),
    uploadedFiles: Array.from(
      { length: faker.number.int({ min: 1, max: 10 }) },
      (_, i) => i + 1
    ).map(() => ({
      id: faker.string.alpha({ length: { min: 10, max: 20 } }),
      name: faker.string.alpha({ length: { min: 10, max: 20 } }),
      size: faker.number.int({ min: 0 }),
      type: faker.string.alpha({ length: { min: 10, max: 20 } }),
    })),
    additionalContent: {
      keyHighlights: faker.string.alpha({ length: { min: 10, max: 20 } }),
      developerInfo: faker.string.alpha({ length: { min: 10, max: 20 } }),
      nearbyLocations: faker.string.alpha({ length: { min: 10, max: 20 } }),
      specialOffers: faker.string.alpha({ length: { min: 10, max: 20 } }),
    },
    enabledToolIds: Array.from(
      { length: faker.number.int({ min: 1, max: 10 }) },
      (_, i) => i + 1
    ).map(() => faker.string.alpha({ length: { min: 10, max: 20 } })),
  },
  sections: faker.helpers.arrayElement([
    Array.from(
      { length: faker.number.int({ min: 1, max: 10 }) },
      (_, i) => i + 1
    ).map(() => ({})),
    null,
  ]),
  savedAt: faker.helpers.arrayElement([
    faker.date.past().toISOString().slice(0, 19) + "Z",
    null,
  ]),
  ...overrideResponse,
})

export const getPublishWebsiteResponseMock = (
  overrideResponse: Partial<Extract<Website, object>> = {}
): Website => ({
  id: faker.string.uuid(),
  name: faker.string.alpha({ length: { min: 10, max: 20 } }),
  domain: faker.string.alpha({ length: { min: 10, max: 20 } }),
  status: faker.helpers.arrayElement(["live", "draft"] as const),
  thumbnailUrl: faker.internet.url(),
  views: faker.number.int({ min: 0 }),
  leads: faker.number.int({ min: 0 }),
  conversionRate: faker.number.float({ fractionDigits: 2 }),
  toolIds: Array.from(
    { length: faker.number.int({ min: 1, max: 10 }) },
    (_, i) => i + 1
  ).map(() => faker.string.alpha({ length: { min: 10, max: 20 } })),
  templateId: faker.string.alpha({ length: { min: 10, max: 20 } }),
  createdAt: faker.date.past().toISOString().slice(0, 19) + "Z",
  updatedAt: faker.date.past().toISOString().slice(0, 19) + "Z",
  publishedAt: faker.helpers.arrayElement([
    faker.date.past().toISOString().slice(0, 19) + "Z",
    null,
  ]),
  ...overrideResponse,
})

export const getListWebsitesMockHandler = (
  overrideResponse?:
    | Website[]
    | ((
        info: Parameters<Parameters<typeof http.get>[1]>[0]
      ) => Promise<Website[]> | Website[]),
  options?: RequestHandlerOptions
) => {
  return http.get(
    "/api/websites",
    async (info: Parameters<Parameters<typeof http.get>[1]>[0]) => {
      return HttpResponse.json(
        overrideResponse !== undefined
          ? typeof overrideResponse === "function"
            ? await overrideResponse(info)
            : overrideResponse
          : getListWebsitesResponseMock(),
        { status: 200 }
      )
    },
    options
  )
}

export const getCreateWebsiteMockHandler = (
  overrideResponse?:
    | Website
    | ((
        info: Parameters<Parameters<typeof http.post>[1]>[0]
      ) => Promise<Website> | Website),
  options?: RequestHandlerOptions
) => {
  return http.post(
    "/api/websites",
    async (info: Parameters<Parameters<typeof http.post>[1]>[0]) => {
      return HttpResponse.json(
        overrideResponse !== undefined
          ? typeof overrideResponse === "function"
            ? await overrideResponse(info)
            : overrideResponse
          : getCreateWebsiteResponseMock(),
        { status: 201 }
      )
    },
    options
  )
}

export const getGetWebsiteMockHandler = (
  overrideResponse?:
    | Website
    | ((
        info: Parameters<Parameters<typeof http.get>[1]>[0]
      ) => Promise<Website> | Website),
  options?: RequestHandlerOptions
) => {
  return http.get(
    "/api/websites/:websiteId",
    async (info: Parameters<Parameters<typeof http.get>[1]>[0]) => {
      return HttpResponse.json(
        overrideResponse !== undefined
          ? typeof overrideResponse === "function"
            ? await overrideResponse(info)
            : overrideResponse
          : getGetWebsiteResponseMock(),
        { status: 200 }
      )
    },
    options
  )
}

export const getGetWebsiteContentMockHandler = (
  overrideResponse?:
    | WebsiteContent
    | ((
        info: Parameters<Parameters<typeof http.get>[1]>[0]
      ) => Promise<WebsiteContent> | WebsiteContent),
  options?: RequestHandlerOptions
) => {
  return http.get(
    "/api/websites/:websiteId/content",
    async (info: Parameters<Parameters<typeof http.get>[1]>[0]) => {
      return HttpResponse.json(
        overrideResponse !== undefined
          ? typeof overrideResponse === "function"
            ? await overrideResponse(info)
            : overrideResponse
          : getGetWebsiteContentResponseMock(),
        { status: 200 }
      )
    },
    options
  )
}

export const getSaveWebsiteContentMockHandler = (
  overrideResponse?:
    | WebsiteContent
    | ((
        info: Parameters<Parameters<typeof http.put>[1]>[0]
      ) => Promise<WebsiteContent> | WebsiteContent),
  options?: RequestHandlerOptions
) => {
  return http.put(
    "/api/websites/:websiteId/content",
    async (info: Parameters<Parameters<typeof http.put>[1]>[0]) => {
      return HttpResponse.json(
        overrideResponse !== undefined
          ? typeof overrideResponse === "function"
            ? await overrideResponse(info)
            : overrideResponse
          : getSaveWebsiteContentResponseMock(),
        { status: 200 }
      )
    },
    options
  )
}

export const getPublishWebsiteMockHandler = (
  overrideResponse?:
    | Website
    | ((
        info: Parameters<Parameters<typeof http.post>[1]>[0]
      ) => Promise<Website> | Website),
  options?: RequestHandlerOptions
) => {
  return http.post(
    "/api/websites/:websiteId/publish",
    async (info: Parameters<Parameters<typeof http.post>[1]>[0]) => {
      return HttpResponse.json(
        overrideResponse !== undefined
          ? typeof overrideResponse === "function"
            ? await overrideResponse(info)
            : overrideResponse
          : getPublishWebsiteResponseMock(),
        { status: 200 }
      )
    },
    options
  )
}
export const getWebsitesMock = () => [
  getListWebsitesMockHandler(),
  getCreateWebsiteMockHandler(),
  getGetWebsiteMockHandler(),
  getGetWebsiteContentMockHandler(),
  getSaveWebsiteContentMockHandler(),
  getPublishWebsiteMockHandler(),
]